Create Assignment 2 in CSE1300: 4 functions
For this assignment I was assigned to create 4 functions that would print out correct results based off what the user inputs.

The first function I created calculates a gross salary by multiplying whatever the user input was by House Rent Allowance(40% of salary), Dearness allowance(20% of salary) and Provident fund(12% of salary). My program was able to take the value of the user input and multiply it by each of these values then add it to the final value and print out the result under Gross Salary. I did this by defining calculate the gross salary by the basic salary and created a stored variable called basic salary.I then created variables for the House Rent Allowance, Dearness Allowance, and provident fund under the variables HRA,DA, and PF. I then set those variable to multiply the basic salary by the values I was given to calculate the allowances. I then created a variable called Gross Salary which equaled the value of the basic salary and the values of all the allowances. I then had the function return the gross salary.

The second function I created asked the user for a base value and would multiply it by the exponent value a user gave. I did this by defining calculating the power by the base and exponent. I created variables base and exponent and had the result variable equal base ** exponent. I then had the function return the result

The third function I created is able to translate degrees in fahrenheit to degrees in celsius. I did this by defining convertingFtoC of fahrenheit. I had the variable celsius equal (the user input of fahrenheit - 32) times 5/9. The function will take the entered value of fahrenheit and convert it celsius with this formula. I then had the function return the result of celsius.

The last function was created to return a total payment. I did this by defining the function calculate payment of the principal amount, interest rate, and time period. I created the variable total payment and set it equal to principal amount times (1 + interest rate/100 ** time period). The function will take all of the entered values and calculate them into the total payment. I then had the function return the total payment.