Add files via upload
I created a mini pirate game using python on Replit. I was given starting values such as crew members, officers, morale, cannons, and treasure. The player is asked to enter their name as captain and is then given 4 choices, either attack a town or ship, or bury and divide the treasure amongst the crew. I created functions adding to the treasure and morale depending on what the player does, if they chose to attack a ship or bury the treasure the morale decreases. Once the crew's morale reaches 0 the game ends.